# app/routes/entities_relationships_system.py
# Auto-generated from OpenAPI. Edits may be overwritten. Put business logic in app/services.
from typing import Any, Optional, List, Dict
from uuid import UUID
from datetime import date, datetime
from fastapi import APIRouter, Depends, Path, Query, Body
from app.schemas.models import *  # noqa: F403,F401

router = APIRouter(tags=["entities:relationships:system"])


@router.delete(
    "/entity-relationships/{id}",
    
    status_code=200,
)
async def delete_entity_relationships__id_(id: UUID,) -> None:
    """
    Delete a specific entity relationship
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.entities_relationships_system.delete_entity_relationships__id_(...)
    raise NotImplementedError("DELETE /entity-relationships/{id} not implemented yet")

@router.get(
    "/entity-relationships",
    
    status_code=200,
)
async def get_entity_relationships() -> None:
    """
    List all entity relationships
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.entities_relationships_system.get_entity_relationships(...)
    raise NotImplementedError("GET /entity-relationships not implemented yet")

@router.patch(
    "/entity-relationships/{id}",
    response_model=EntityRelationship,
    status_code=200,
)
async def patch_entity_relationships__id_(id: UUID,body: EntityRelationshipInput = Body(...),) -> EntityRelationship:
    """
    Update a specific entity relationship
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.entities_relationships_system.patch_entity_relationships__id_(...)
    raise NotImplementedError("PATCH /entity-relationships/{id} not implemented yet")

@router.post(
    "/entity-relationships",
    response_model=EntityRelationship,
    status_code=201,
)
async def post_entity_relationships(body: EntityRelationshipInput = Body(...),) -> EntityRelationship:
    """
    Create a new entity relationship
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.entities_relationships_system.post_entity_relationships(...)
    raise NotImplementedError("POST /entity-relationships not implemented yet")