# openapi.yaml
openapi: 3.0.4
info:
  description: |
    The **Tag Management System API** provides RESTful access to all tag-related operations—creation, updates, deletion, linking, and querying of tags and their relationships.
    It also supports metadata assignment for tagged entities such as images, documents, or video stills.

    ## Principles
    - **RESTful architecture** — resource-based URIs and standard HTTP verbs (GET, POST, PATCH, DELETE)
    - **JSON payloads** — all requests and responses use JSON
    - **Frontend-controlled authentication** — only frontend clients (CLI, UI) interact with the API through validated tokens (except for development/testing)
  title: Tag Management System API
  version: 0.1.1
servers:
- description: Local development
  url: http://localhost:8000/api/v1
...
paths:
...
  /entity-relationships:
    get:
      description: "Retrieve a system-wide list of all entity-to-entity relationships,\
        \ regardless of source or target."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_entities__id__relationships_get_200_response"
          description: List of entity relationships
      summary: List all entity relationships
      tags:
      - entities:relationships:system
    post:
      description: "Create a new relationship between two entities, defining their\
        \ linkage type and direction."
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityRelationshipInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityRelationship"
          description: Relationship created
      summary: Create a new entity relationship
      tags:
      - entities:relationships:system
  /entity-relationships/{id}:
    delete:
      description: Permanently remove an entity-to-entity relationship record from
        the system.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Relationship deleted
      summary: Delete a specific entity relationship
      tags:
      - entities:relationships:system
    patch:
      description: Update the relationship type or description for a specific relationship
        record.
      parameters:
      - description: Unique identifier for the relationship record
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityRelationshipInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityRelationship"
          description: Relationship updated
      summary: Update a specific entity relationship
      tags:
      - entities:relationships:system
...
components:
...
  schemas:
...
    EntityRelationship:
      description: |
        Directionality semantics:
        - entity_a_id — source/primary entity
        - entity_b_id — alternate/derived entity
      example:
        relationship_type_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        entity_b_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        entity_a_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        entity_a_id:
          format: uuid
          title: entity_a_id
          type: string
        entity_b_id:
          format: uuid
          title: entity_b_id
          type: string
        relationship_type_id:
          format: uuid
          title: relationship_type_id
          type: string
      title: EntityRelationship
      type: object
    EntityRelationshipInput:
      example:
        relationship_type_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        entity_b_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        entity_b_id:
          format: uuid
          title: entity_b_id
          type: string
        relationship_type_id:
          format: uuid
          title: relationship_type_id
          type: string
      title: EntityRelationshipInput
      type: object
...
    _tags__id__delete_200_response:
      example:
        status: deleted
      properties:
        status:
          example: deleted
          title: status
          type: string
      title: _tags__id__delete_200_response
      type: object
...
    _entities__id__relationships_get_200_response:
      example:
        relationships:
        - relationship_type_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          entity_b_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          entity_a_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - relationship_type_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          entity_b_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          entity_a_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        relationships:
          items:
            $ref: "#/components/schemas/EntityRelationship"
          title: relationships
          type: array
      title: _entities__id__relationships_get_200_response
      type: object
...